requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.1
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
botframework-connector==4.15.0
botframework-streaming==4.15.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
confluent-kafka==2.3.0
cryptography==42.0.5
exceptiongroup==1.2.1
fastapi==0.110.2
fastavro==1.9.4
frozenlist==1.4.1
h11==0.14.0
idna==3.7
isodate==0.6.1
jsonpickle==1.4.2
kafka-python==2.0.2
msal==1.28.0
msrest==0.7.1
multidict==6.0.5
oauthlib==3.2.2
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
urllib3==1.26.18
uvicorn==0.29.0
yarl==1.9.4